// frontend/src/api/models/order-item.ts
/* tslint:disable */
/* eslint-disable */
import { Product } from '../models/product';
export interface OrderItem {
  complements?: Array<number>;
  complements_data: Array<Product>;
  id: number;
  note?: null | string;
  order: number;
  product: number;
  product_data: Product;
  quantity?: number;
}